+++ packages/e2e-tests/inspector-kit/vite.config.js
@@ -1,0 +1,17 @@
+// eslint-disable-next-line node/no-missing-import
+import { sveltekit } from '@sveltejs/kit/vite';
+
+/** @type {import('vite').UserConfig} */
+const config = {
+	server: {
+		watch: {
+			// During tests we edit the files too fast and sometimes chokidar
+			// misses change events, so enforce polling for consistency
+			usePolling: true,
+			interval: 100
+		}
+	},
+	plugins: [sveltekit()]
+};
+
+export default config;