+++ packages/e2e-tests/vite-ssr/server.js
@@ -30,9 +30,14 @@ async function createServer(root = process.cwd(), isProd = process.env.NODE_ENV 
 	if (!isProd) {
 		const inlineCfg = {
 			root,
+			appType: 'custom',
 			server: {
 				middlewareMode: true,
-				port
+				port,
+				strictPort: true,
+				hmr: {
+					port: port + 25000
+				}
 			}
 		};
 		// @ts-ignore